perf(web): AgeVerificationWrapper をオーバーレイ方式に変更してコンテンツブロッキングを解消（SPR-7）
レイアウトを <AgeVerificationProvider> 直下に SessionProvider + コンテンツを
直置きし、年齢確認モーダルを fixed inset-0 z-50 のオーバーレイ
(<AgeVerificationOverlay>) として後出しに変更。Wrapper / Gate / 未使用と
なった server-side bot-detection を削除。コンテンツが localStorage チェック
完了前にも描画されるようになり、初回訪問時の FCP/LCP を改善する。

- 旧 Gate の挙動（18歳未満分岐 + 3 秒後リダイレクト、30 日記憶、bot UA 通過
  による SEO 維持）は overlay 側で保持
- R18 フィルタは元から showR18Content ベースでコンテンツ層に分散実装されて
  おり、wrapper 除去による機能影響なし
- オーバーレイ背景は suzuka-gradient ユーティリティクラスを使用
  （Tailwind v4 では from-X / via-X / to-X が手動定義のテーマトークンを解決
  できず透明になるため）
- 表示中は body overflow を hidden にしてスクロールロック
